+++ Data/Scripts/RealEnergy/Behaviors/HeatVentComponent.cs
@@ -1,0 +1,96 @@
+using System.Text;
+using Sandbox.Common.ObjectBuilders;
+using Sandbox.ModAPI;
+using SpaceEngineers.Game.ModAPI;
+using VRage.Game.Components;
+using VRage.Game.ModAPI;
+using VRage.ObjectBuilders;
+
+namespace TSUT.HeatManagement
+{
+    [MyEntityComponentDescriptor(typeof(MyObjectBuilder_HeatVentBlock), false)]
+    public class HeatVentComponent : AHeatGameLogicComponent
+    {
+        private static bool _controlsRegistered = false;
+
+        private IMyHeatVent _vent;
+
+        public override void UpdateOnceBeforeFrame()
+        {
+            _vent = (IMyHeatVent)Entity;
+            _vent.AppendingCustomInfo += AppendVentHeatInfo;
+
+            base.UpdateOnceBeforeFrame();
+
+            RegisterTerminalControls();
+        }
+
+        public override void Cleanup()
+        {
+            if (Entity != null)
+                _vent.AppendingCustomInfo -= AppendVentHeatInfo;
+        }
+
+        public override float GetHeatChange(float deltaTime)
+        {
+            if (_vent == null)
+                return 0f;
+
+            float change = HeatSession.Api.Utils.GetAmbientHeatLoss(_vent, deltaTime);
+            if (_vent.IsWorking)
+                change += HeatSession.Api.Utils.GetActiveHeatVentLoss(_vent, deltaTime);
+
+            return -change;
+        }
+
+        public override void SpreadHeat(float deltaTime)
+        {
+            SpreadHeatStandard(_vent, deltaTime);
+        }
+
+        public override void ReactOnNewHeat(float heat)
+        {
+            _vent.RefreshCustomInfo();
+            _vent.SetDetailedInfoDirty();
+        }
+
+        private void AppendVentHeatInfo(IMyTerminalBlock block, StringBuilder builder)
+        {
+            float ownThermalCapacity = HeatSession.Api.Utils.GetThermalCapacity(block);
+
+            var neighborStringBuilder = new StringBuilder();
+            float cumulativeNeighborHeatChange;
+            float cumulativeNetworkHeatChange;
+            AddNeighborAndNetworksInfo(_vent, neighborStringBuilder, out cumulativeNeighborHeatChange, out cumulativeNetworkHeatChange);
+
+            float heatChange = GetHeatChange(1f) - cumulativeNeighborHeatChange - cumulativeNetworkHeatChange;
+            float airDensity = HeatSession.Api.Utils.GetAirDensity(_vent);
+
+            builder.AppendLine($"--- Heat Management ---");
+            builder.AppendLine($"Temperature: {HeatSession.Api.Utils.GetHeat(block):F2} °C");
+            string heatStatus = heatChange > 0 ? "Heating" : heatChange < -0.01 ? "Cooling" : "Stable";
+            builder.AppendLine($"Thermal Status: {heatStatus}");
+            builder.AppendLine($"Net Heat Change: {heatChange:+0.00;-0.00;0.00} °C/s");
+            string exchangeMode = _vent.IsWorking ? "Active" : "Passive";
+            builder.AppendLine($"Exchange Mode: {exchangeMode}");
+            builder.AppendLine($"Thermal Capacity: {ownThermalCapacity / 1000000:F1} MJ/°C");
+            builder.AppendLine($"Ambient temp: {HeatSession.Api.Utils.CalculateAmbientTemperature(block):F1} °C");
+            builder.AppendLine($"Air density: {airDensity * 100:F1} %");
+            float windSpeed = HeatSession.Api.Utils.GetBlockWindSpeed(block);
+            builder.AppendLine($"Wind Speed: {windSpeed:F2} m/s");
+            builder.AppendLine($"------");
+            builder.AppendLine("");
+            builder.AppendLine("Heat Sources:");
+            builder.AppendLine($"  Air Exchange: {GetHeatChange(1):+0.00;-0.00;0.00} °C/s");
+            builder.Append(neighborStringBuilder);
+        }
+
+        private static void RegisterTerminalControls()
+        {
+            if (_controlsRegistered) return;
+            _controlsRegistered = true;
+
+            HeatSession.Api.Utils.TryRegister<IMyHeatVent>();
+        }
+    }
+}
