+++ Data/Scripts/RealEnergy/Session.cs
@@ -61,9 +61,7 @@ namespace TSUT.HeatManagement
             Instance = this;
             MyLog.Default.WriteLine($"[HeatManagement] HeatSession instance created.");
 
-            _heatApi.Registry.RegisterHeatBehaviorFactory(new ExhaustHeatManagerFactory());
             _heatApi.Registry.RegisterHeatBehaviorFactory(new HeatPipeManagerFactory());
-            _heatApi.Registry.RegisterHeatBehaviorFactory(new HeatVentManagerFactory());
 
             MyAPIGateway.Utilities.RegisterMessageHandler(HmsApi.HeatProviderMesageId, OnHeatProviderRegister);
             var shareable = ConvertApiToShareable(_heatApi);
